+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Track user IP on Flask
 - Make user IP tracking on Bottle and Pyramid use the same algorithm as other
   integrations, checking for the `client-ip` header
+- Add support to Falcon integration for tracking request queue timing
 
 ### Fixed